+++ template_new_cli/RELEASES.md
@@ -10,4 +10,6 @@ The TODO section is part of the [README.md](https://github.com/bestia-dev/cargo_
 
 ## Unreleased
 
-## Version 0.0.0
+## Version 0.0.1
+
+- Rust project created with `cargo auto new_cli cargo_auto_template_new_cli`